requirements.txt
pyyaml>=6.0.1
hydra-core>=1.3.2
omegaconf>=2.3.0
pydantic>=2.5
pandas>=2.2
numpy>=1.26
scikit-learn>=1.3
imbalanced-learn>=0.12
ydata-profiling>=4.6
umap-learn>=0.5
tpot>=0.12
matplotlib>=3.8
jinja2>=3.1
streamlit>=1.36